backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
transformers
torch
vaderSentiment
python-dotenv
pandas
scikit-learn
